Add exercises 4.19

+++ part4/controllers/blog.js
@@ -50,8 +50,24 @@ blogsRouter.post('/', userExtractor, async (request, response) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
   try {
+    const decodedToken = request.user
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response.status(404).end()
+    }
+
+    if (!blog.user || blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: 'unauthorized' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   } catch (error) {
